skip config reference check in sync (fixes #57)
Git and other clients such as SourceTree will rewrite .git/config
in the modules when checking out a branch tracking a remote branch.
The rewrite appears to be a remove then recreate operation, which
efectively removes the .git/config symlink.

As a consequence, repo sync will complain in project._CheckDirReference()
will detect the src and dst for 'config' to be different and assume there
are local changes which need not be overwritten/lost.

As a workaround, from the sync context we skip the check on the config file
to be correctly linked.

This is probabaly NOT the Right Way(TM) to fix this, but is enough to
allow the sync to work in a way which does not seem to be catastrophic
or obviously broken.

Probably the correct solution would be to remove the .repo config and
replace and re-symlink it with the one in the module itself.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>